Reorganize main folders
