+++ torchqdynamics/smesolve/__init__.py
@@ -1,0 +1,1 @@
+from .smesolve import smesolve
